all tests pass for new floats

+++ samples/boat1/cmd.go
@@ -59,7 +59,7 @@ func (b *Boat) Spin(ctx context.Context, angleDeg float64, degsPerSec float64, b
 	return math.NaN(), fmt.Errorf("boat can't spin yet")
 }
 
-func (b *Boat) WidthMillis(ctx context.Context) (int, error) {
+func (b *Boat) WidthMillis(ctx context.Context) (float64, error) {
 	return 1, nil
 }
